requirements.txt
loguru==0.5.1
numpy==1.19.0
pandas==1.0.5
pytorch-ignite==0.3.0
scipy==1.4.1
tensorboard==2.1.0
torch==1.3.1
torchvision==0.4.2